Addition of new page to calculate speed required for most probable nanosheet size.
